🐛 fix: 修复代码走查发现的多个问题
- 修复整数溢出导致负数索引 (FastDFSConnection, RoundRobinStorageSelector)
- 添加 DNS 解析超时控制 (30秒)
- 修复 DisposeAsync 返回同步完成的 ValueTask
- 修复 Dispose 吞掉异常，改为记录日志
- IConnectionPool/IConnectionPoolProvider 添加 IAsyncDisposable
- 重构 ConnectionPool Timer 回调为独立方法
- NetworkTimeout 验证改为兼容处理，不再抛异常
- ByteExtensions.Fill 填充0时使用 Array.Clear 优化

+++ src/FastDFS.Client/Storage/RoundRobinStorageSelector.cs
@@ -34,7 +34,8 @@ namespace FastDFS.Client.Storage
                 return servers[0];
 
             // Thread-safe increment and wrap around
-            int index = Interlocked.Increment(ref _currentIndex) % servers.Count;
+            // Use unsigned modulo to avoid negative index after int overflow (~2 billion increments)
+            int index = (int)((uint)Interlocked.Increment(ref _currentIndex) % servers.Count);
             return servers[index];
         }
     }
